+++ jabgui/src/main/java/org/jabref/gui/preferences/PreferencesDialogViewModel.java
@@ -26,6 +26,7 @@ import org.jabref.gui.preferences.export.ExportTab;
 import org.jabref.gui.preferences.external.ExternalTab;
 import org.jabref.gui.preferences.externalfiletypes.ExternalFileTypesTab;
 import org.jabref.gui.preferences.general.GeneralTab;
+import org.jabref.gui.preferences.git.GitTab;
 import org.jabref.gui.preferences.groups.GroupsTab;
 import org.jabref.gui.preferences.journals.JournalAbbreviationsTab;
 import org.jabref.gui.preferences.keybindings.KeyBindingsTab;
@@ -96,7 +97,8 @@ public class PreferencesDialogViewModel extends AbstractViewModel {
                 new XmpPrivacyTab(),
                 new CustomImporterTab(),
                 new CustomExporterTab(),
-                new NetworkTab()
+                new NetworkTab(),
+                new GitTab()
         );
     }